% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSliver.R
\name{deSliver}
\alias{deSliver}
\title{A function for merging sliver polygons into non-sliver neighbours.
The threshold is applied to the area of the multipolygon object, not each
individual polygon. Non-sliver polygons keep their original attributesl
Inteded to be used when it is important to retain the original extent of an
area while removing sliver polygons}
\usage{
deSliver(x, threshold)
}
\arguments{
\item{x}{A spatialPolygonsDataFrame or sf object}

\item{threshold}{the minimum area below which a polygon is considered a sliver}
}
\value{
an object of class sf or Spatial with sliver polygons
merged to their nearest valid neighbour.
}
\description{
A function for merging sliver polygons into non-sliver neighbours.
The threshold is applied to the area of the multipolygon object, not each
individual polygon. Non-sliver polygons keep their original attributesl
Inteded to be used when it is important to retain the original extent of an
area while removing sliver polygons
}
\examples{
deSliver(x = intersectedPolygons, threshold = 500)
}
\keyword{intersect}
\keyword{polygons}
\keyword{sliver}
