% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMask.R
\name{multiMask}
\alias{multiMask}
\title{A function for masking multiple values at once
This function is simply lapply with mask}
\usage{
multiMask(maskValues, ...)
}
\description{
A function for masking multiple values at once
This function is simply lapply with mask
}
\examples{
multiMask(aRaster, maskValues = c(5,10))
}
\keyword{mask}
\keyword{raster}
