% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerBuffer.R
\name{outerBuffer}
\alias{outerBuffer}
\title{DRAW CONVEX HILL AROUND POLYGONS -----------------------
draws a convex hull aroung vertice points of a polygon shape file.
x must be a SpatialPolygons, or SpatialPolygonsDF
Courtesy of Ceres Barros}
\usage{
outerBuffer(x)
}
\arguments{
\item{x}{A SpatialPolygons or SpatialPolygonsDataFrame object}
}
\description{
DRAW CONVEX HILL AROUND POLYGONS -----------------------
draws a convex hull aroung vertice points of a polygon shape file.
x must be a SpatialPolygons, or SpatialPolygonsDF
Courtesy of Ceres Barros
}
\examples{
outerBuffer()
}
\keyword{buffer}
\keyword{merge}
\keyword{polygons}
