% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unSliver.R
\name{unSliver}
\alias{unSliver}
\title{A function for merging sliver polygons into non-sliver neighbours.}
\usage{
unSliver(x, threshold)
}
\arguments{
\item{x}{A spatialPolygonsDataFrame or sf object}

\item{threshold}{the minimum area below which a polygon is considered a sliver}
}
\value{
an object of class sf or Spatial with sliver polygons
merged to their nearest valid neighbour.
}
\description{
A function for merging sliver polygons into non-sliver neighbours.
}
\details{
The threshold is applied to the area of the multipolygon object, not each
individual polygon. Non-sliver polygons keep their original attributes
}
\examples{
unSliver(x = intersectedPolygons, threshold = 500)
}
\keyword{intersect}
\keyword{polygons}
\keyword{sliver}
